refactor: naming and moving things around
